% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_context.R
\name{add_context}
\alias{add_context}
\title{Add context to search results}
\usage{
add_context(gramx, df, form = "token", tag = "xpos", highlight = NULL)
}
\arguments{
\item{gramx}{output from find_gramx()}

\item{df}{Annotion dataframe}

\item{form}{A character string}

\item{tag}{A character string}

\item{highlight}{A boolean}
}
\value{
A data frame
}
\description{
Add context to search results
}
