% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biber_tags.R
\name{biber_tags}
\alias{biber_tags}
\title{Re-code PoS tags, per Biber (1991)}
\usage{
biber_tags(token, tag)
}
\arguments{
\item{x}{Annotion dataframe}
}
\value{
A data frame
}
\description{
Re-code PoS tags, per Biber (1991)
}
