% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_inline.R
\name{build_inline}
\alias{build_inline}
\title{Build inline tif for search gramx}
\usage{
build_inline(df, form, tag)
}
\arguments{
\item{df}{A data frame}

\item{form}{vector}

\item{tag}{vector}
}
\value{
A data frame
}
\description{
Build inline tif for search gramx
}
