% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lexical_features.R
\name{count_lexical_features}
\alias{count_lexical_features}
\title{Count ~lexical features from Biber (1991)}
\usage{
count_lexical_features(x)
}
\arguments{
\item{x}{Annotion dataframe}
}
\value{
A data frame
}
\description{
Count ~lexical features from Biber (1991)
}
