% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gramx.R
\name{find_gramx}
\alias{find_gramx}
\title{Extract gramx from tif}
\usage{
find_gramx(search, mapping = textsearch::mapping_generic, tif)
}
\arguments{
\item{search}{char string}

\item{mapping}{A list}

\item{tif}{a tif}
}
\value{
A data frame
}
\description{
Extract gramx from tif
}
