% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lex.R
\name{find_lex}
\alias{find_lex}
\title{Identify lexical patterns in text.}
\usage{
find_lex(query, text, doc_id, window = 20L, highlight = c("<", ">"))
}
\arguments{
\item{query}{a character vector}

\item{text}{a character vector}

\item{doc_id}{a character vector}

\item{window}{An integer}

\item{highlight}{a character vector}
}
\value{
A data frame
}
\description{
Identify lexical patterns in text.
}
