% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gramx.R
\name{recode_gramx}
\alias{recode_gramx}
\title{Recode gramx in df}
\usage{
recode_gramx(
  gramx,
  df,
  form = "token",
  tag = "xpos",
  recode_col = "xpos",
  recode_cat = "cx1"
)
}
\arguments{
\item{gramx}{A DF}

\item{df}{An annotated corpus df}

\item{recode_col}{A string}

\item{recode_cat}{A string}
}
\value{
A data frame
}
\description{
Recode gramx in df
}
