% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_query.R
\name{translate_query}
\alias{translate_query}
\title{Build regex query}
\usage{
translate_query(x, mapping = mapping)
}
\arguments{
\item{x}{a string}

\item{mapping}{a list}
}
\value{
A string
}
\description{
Build regex query
}
