% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gramx.R
\name{find_gramx}
\alias{find_gramx}
\title{Extract gramx from tif}
\usage{
find_gramx(tif, query)
}
\arguments{
\item{tif}{a tif}

\item{search}{char string}
}
\value{
A data frame
}
\description{
Extract gramx from tif
}
