% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gramx.R
\name{recode_gramx}
\alias{recode_gramx}
\title{Recode gramx in df}
\usage{
recode_gramx(
  gramx,
  df,
  form = "token",
  tag = "xpos",
  col = "xpos",
  new_cat = "cx1",
  renumber = T
)
}
\arguments{
\item{gramx}{A DF}

\item{df}{An annotated corpus df}

\item{col}{A string}

\item{new_cat}{A string}

\item{renumber}{A boolean}
}
\value{
A data frame
}
\description{
Recode gramx in df
}
